develop/python/key-tracker/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
alembic>=1.11.0
pyyaml>=6.0
python-multipart